+++ src/main/java/com/telepathicgrunt/the_bumblezone/entities/BeeAggression.java
@@ -10,7 +10,9 @@ import com.telepathicgrunt.the_bumblezone.modinit.BzEffects;
 import com.telepathicgrunt.the_bumblezone.modinit.BzTags;
 import net.fabricmc.fabric.api.event.lifecycle.v1.ServerLifecycleEvents;
 import net.fabricmc.fabric.api.event.player.PlayerBlockBreakEvents;
+import net.minecraft.ChatFormatting;
 import net.minecraft.core.Registry;
+import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceKey;
 import net.minecraft.server.MinecraftServer;
 import net.minecraft.server.level.ServerLevel;
@@ -70,7 +72,7 @@ public class BeeAggression {
     //if player mines an angerable tagged block, bees gets very mad...
     public static void blockBreakAnger(Player player, Block block) {
         if (player instanceof ServerPlayer serverPlayer &&
-            EssenceOfTheBees.hasEssence(serverPlayer) &&
+            !EssenceOfTheBees.hasEssence(serverPlayer) &&
             block.defaultBlockState().is(BzTags.WRATH_ACTIVATING_BLOCKS_WHEN_MINED))
         {
             angerBees(player);
@@ -80,7 +82,7 @@ public class BeeAggression {
     //if player picks up an angerable tagged item, bees gets very mad...
     public static void itemPickupAnger(Player player, Item item) {
         if (player instanceof ServerPlayer serverPlayer &&
-            EssenceOfTheBees.hasEssence(serverPlayer) &&
+            !EssenceOfTheBees.hasEssence(serverPlayer) &&
             item.getDefaultInstance().is(BzTags.WRATH_ACTIVATING_ITEMS_WHEN_PICKED_UP))
         {
             angerBees(player);
@@ -97,7 +99,11 @@ public class BeeAggression {
                 !player.isSpectator())
         {
             if(!player.hasEffect(BzEffects.PROTECTION_OF_THE_HIVE)) {
-                //Bumblezone.LOGGER.log(Level.INFO, "ANGRY BEES");
+                if (player instanceof ServerPlayer) {
+                    Component message = Component.translatable("system.the_bumblezone.no_protection").withStyle(ChatFormatting.BOLD).withStyle(ChatFormatting.RED);
+                    player.displayClientMessage(message, true);
+                }
+
                 player.addEffect(new MobEffectInstance(
                         BzEffects.WRATH_OF_THE_HIVE,
                         BzConfig.howLongWrathOfTheHiveLasts,
@@ -105,6 +111,9 @@ public class BeeAggression {
                         false,
                         BzConfig.showWrathOfTheHiveParticles,
                         true));
+            }
+            else if (player instanceof ServerPlayer serverPlayer) {
+                BzCriterias.HONEY_PERMISSION_TRIGGER.trigger(serverPlayer);
             }
         }
     }
@@ -240,6 +249,11 @@ public class BeeAggression {
         StructureManager structureManager = ((ServerLevel)serverPlayer.level).structureManager();
         if (structureManager.getStructureWithPieceAt(serverPlayer.blockPosition(), BzTags.WRATH_CAUSING).isValid()) {
             if (!serverPlayer.hasEffect(BzEffects.PROTECTION_OF_THE_HIVE)) {
+                if (!serverPlayer.hasEffect(BzEffects.WRATH_OF_THE_HIVE)) {
+                    Component message = Component.translatable("system.the_bumblezone.no_protection").withStyle(ChatFormatting.BOLD).withStyle(ChatFormatting.RED);
+                    serverPlayer.displayClientMessage(message, true);
+                }
+
                 serverPlayer.addEffect(new MobEffectInstance(
                         BzEffects.WRATH_OF_THE_HIVE,
                         BzConfig.howLongWrathOfTheHiveLasts,