+++ CHANGELOG.md
@@ -3,6 +3,15 @@
 ##### Effects:
 Getting preventable Wrath of the Hive effect will display a message saying you lacked Protection of the Hive to prevent it.
 
+##### Advancements:
+Added two new advancements to help introduce Protection of the Hive effect.
+
+Fixed honey_bucket_brood advancement triggering when feeding Honeycomb Brood Blocks anything. 
+ Now it should only trigger when feeding Honey Buckets/Royal Jelly Buckets to Honeycomb Brood Blocks.
+
+##### Essense:
+Fixed Essence of the Bees effect not preventing wrath effect when mining Honey Blocks.
+
 ##### Config:
 Fixed some config comments